# Improve BytesRef creation from String