### Description

<!--
If this is your first contribution to Lucene, please make sure you have reviewed the contribution guide.
https://github.com/apache/lucene/blob/main/CONTRIBUTING.md
-->

When creating a `BytesRef` from a String, we can leverage `String::getBytes(StandardCharset.UTF_8)` to avoid 3x memory allocation overhead when the `String` is already `UTF-8` and when JDK provides compressed String representation of `byte[]` this becomes a single array copy rather than more complex
`org.apache.lucene.util.UnicodeUtil#UTF16toUTF8` logic.

#### Before
```
Benchmark                                         (size)  Mode  Cnt      Score      Error   Units
BytesRefBenchmark.newBytesRef                         10  avgt   15     13.723 ±    0.058   ns/op
BytesRefBenchmark.newBytesRef:gc.alloc.rate.norm      10  avgt   15     72.000 ±    0.001    B/op
BytesRefBenchmark.newBytesRef                        100  avgt   15    241.423 ±   58.112   ns/op
BytesRefBenchmark.newBytesRef:gc.alloc.rate.norm     100  avgt   15    616.003 ±    0.001    B/op
BytesRefBenchmark.newBytesRef                       1000  avgt   15   2701.883 ±  603.846   ns/op
BytesRefBenchmark.newBytesRef:gc.alloc.rate.norm    1000  avgt   15   6400.031 ±    0.007    B/op
```


#### After
```
Benchmark                                         (size)  Mode  Cnt      Score      Error   Units
BytesRefBenchmark.bytesRefsOf                         10  avgt   15      5.820 ±    0.055   ns/op
BytesRefBenchmark.bytesRefsOf:gc.alloc.rate.norm      10  avgt   15     56.000 ±    0.001    B/op
BytesRefBenchmark.bytesRefsOf                        100  avgt   15     14.298 ±    0.213   ns/op
BytesRefBenchmark.bytesRefsOf:gc.alloc.rate.norm     100  avgt   15    232.000 ±    0.001    B/op
BytesRefBenchmark.bytesRefsOf                       1000  avgt   15     93.568 ±   10.824   ns/op
BytesRefBenchmark.bytesRefsOf:gc.alloc.rate.norm    1000  avgt   15   2160.001 ±    0.001    B/op
```