@msfroh 2025-05-16
That's a nice improvement! 

Presumably this kicks in when we have [compact strings](https://openjdk.org/jeps/254)? Do you have a JMH benchmark with non-ISO-8859-1/Latin-1 characters?