@schlosna 2025-05-16
> Presumably this kicks in when we have [compact strings](https://openjdk.org/jeps/254)? Do you have a JMH benchmark with non-ISO-8859-1/Latin-1 characters?

Yes, this leverages compact strings for Latin-1 for the best case scenario here. I'll throw together a more rigorous benchmark including non-Latin1.