@rmuir 2025-05-17
There's intentionally not a `BytesRef(String)` constructor, as there have been issues before.