@rmuir 2025-05-17
git log tells the story. A quick summary:

* conflicting use cases such as usages as a reference vs stringbuilder/string type uses
* ownership issues and API problems resulting from that.

The use-cases were torn apart and the code cleaned up and bytesrefbuilder made separate. we shouldn't go backwards.

this is basically a reference class, intended to support pointing to byte[]s held elsewhere. like a pointer in c lol.