@schlosna 2025-05-20
> > In #12071 these is mention [#12071 (comment)](https://github.com/apache/lucene/issues/12071#issuecomment-1379313710) of using the vector APIs to speed up UnicodeUtil conversions. Has any of that been approached in lucene yet? I imagine a SIMD or SWAR approach like https://lemire.me/blog/2018/05/16/validating-utf-8-strings-using-as-little-as-0-7-cycles-per-byte/ & https://lemire.me/blog/2023/02/16/computing-the-utf-8-size-of-a-latin-1-string-quickly-avx-edition/ would be worth investigating.
> 
> If you want to speed something up, pay no mind to java.lang.String as it isn't involved in the indexing anyway. Look here which is what the indexer calls jazillions of times https://github.com/apache/lucene/blob/main/lucene/core/src/java/org/apache/lucene/analysis/tokenattributes/CharTermAttributeImpl.java#L91-L95
> 
> it is just char[]->byte[] conversion, with the same reused char[] and byte[], there aren't allocations or strings or any of that.

Thanks, I'm going to close out this PR for now. When I have a chance I will take a look at optimizing the `char[]` to `byte[]` conversions via `jdk.incubator.vector.ShortVector.fromCharArray`.