@rmuir 2025-05-21
yeah it is tricky, since the 'strings' indexed in search are usually small: words. for a lot of natural languages average word length is already small (e.g. english: ~5), and often in search people are shortening that even more (with stemmers etc).